Add docx support
